% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{scbenchr_add_percent_mt}
\alias{scbenchr_add_percent_mt}
\title{Add percent.mt to a Seurat object}
\usage{
scbenchr_add_percent_mt(seu, pattern = "^MT-")
}
\arguments{
\item{seu}{Seurat object.}

\item{pattern}{Regex to identify mitochondrial genes.}
}
\description{
Add percent.mt to a Seurat object
}
