% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_layers.R
\name{scbenchr_counts_layers}
\alias{scbenchr_counts_layers}
\title{Get counts layers for an Assay5 RNA assay}
\usage{
scbenchr_counts_layers(seu, assay = "RNA")
}
\description{
Get counts layers for an Assay5 RNA assay
}
\keyword{internal}
