% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{scbenchr_default_base_dir}
\alias{scbenchr_default_base_dir}
\title{Default base directory for scBenchR runs (Windows)}
\usage{
scbenchr_default_base_dir()
}
\description{
Default base directory for scBenchR runs (Windows)
}
\keyword{internal}
