% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_rpca.R
\name{scbenchr_export_merged_metadata}
\alias{scbenchr_export_merged_metadata}
\title{Export merged metadata (cell-level) for paper step}
\usage{
scbenchr_export_merged_metadata(seu, out_dir, filename = "merged_metadata.csv")
}
\arguments{
\item{seu}{Integrated Seurat object.}

\item{out_dir}{Output directory.}

\item{filename}{Output filename in integrated/ subfolder.}
}
\description{
Export merged metadata (cell-level) for paper step
}
