% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{scbenchr_init_outdirs}
\alias{scbenchr_init_outdirs}
\title{Create output directory structure}
\usage{
scbenchr_init_outdirs(out_dir)
}
\arguments{
\item{out_dir}{Output directory.}
}
\value{
The normalized output directory path (invisibly).
}
\description{
Create output directory structure
}
