% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_rpca.R
\name{scbenchr_integrate_rpca}
\alias{scbenchr_integrate_rpca}
\title{RPCA integration of two Seurat objects (serial)}
\usage{
scbenchr_integrate_rpca(seu_small, seu_large, npcs = 30, resolution = 0.6)
}
\arguments{
\item{seu_small}{Filtered small object (counts state).}

\item{seu_large}{Filtered large object (counts state).}

\item{npcs}{Number of PCs / dims.}

\item{resolution}{Clustering resolution.}
}
\value{
Integrated Seurat object (DefaultAssay integrated).
}
\description{
RPCA integration of two Seurat objects (serial)
}
