% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers_step4.R
\name{scbenchr_merge_counts_layers}
\alias{scbenchr_merge_counts_layers}
\title{Merge multi-layer counts into a single sparse matrix (for safe DE)}
\usage{
scbenchr_merge_counts_layers(seu, assay = "RNA")
}
\arguments{
\item{seu}{Integrated object with Assay5 RNA having counts.* layers.}

\item{assay}{Assay name.}
}
\value{
Sparse dgCMatrix of merged counts.
}
\description{
Merge multi-layer counts into a single sparse matrix (for safe DE)
}
