% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{scbenchr_qc_filter}
\alias{scbenchr_qc_filter}
\title{Filter Seurat object using QC thresholds}
\usage{
scbenchr_qc_filter(seu, min_features, max_features, min_counts, max_mt)
}
\arguments{
\item{seu}{Seurat object.}

\item{min_features}{Minimum nFeature_RNA.}

\item{max_features}{Maximum nFeature_RNA.}

\item{min_counts}{Minimum nCount_RNA.}

\item{max_mt}{Maximum percent.mt.}
}
\description{
Filter Seurat object using QC thresholds
}
