% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{scbenchr_qc_summary}
\alias{scbenchr_qc_summary}
\title{QC summary (printed)}
\usage{
scbenchr_qc_summary(seu, name = "object")
}
\arguments{
\item{seu}{Seurat object.}

\item{name}{Label printed in output.}
}
\description{
QC summary (printed)
}
