% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_readers.R
\name{scbenchr_read_expMat_safe}
\alias{scbenchr_read_expMat_safe}
\title{Read expression matrix safely (fix duplicated genes/cells)}
\usage{
scbenchr_read_expMat_safe(path, cell_prefix)
}
\arguments{
\item{path}{Path to a gzipped or plain text matrix with first column genes and remaining columns cells.}

\item{cell_prefix}{Prefix to guarantee cell name uniqueness across inputs.}
}
\value{
A numeric matrix with unique rownames and colnames.
}
\description{
Read expression matrix safely (fix duplicated genes/cells)
}
