% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3_serial.R
\name{scbenchr_run_serial_step3}
\alias{scbenchr_run_serial_step3}
\title{Run serial Seurat pipeline (Normalize -> HVG -> Scale -> PCA -> Neighbors -> Clusters -> UMAP)}
\usage{
scbenchr_run_serial_step3(seu, prefix = "DATA", npcs = 30, resolution = 0.6)
}
\arguments{
\item{seu}{Seurat object.}

\item{prefix}{String prefix for timing step names.}

\item{npcs}{Number of PCs.}

\item{resolution}{Clustering resolution.}
}
\value{
list(seu=SeuratObject, timing=data.frame)
}
\description{
Run serial Seurat pipeline (Normalize -> HVG -> Scale -> PCA -> Neighbors -> Clusters -> UMAP)
}
