% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_timing.R
\name{scbenchr_time_it}
\alias{scbenchr_time_it}
\title{Time and return a standardized timing row}
\usage{
scbenchr_time_it(step, expr)
}
\description{
Time and return a standardized timing row
}
\keyword{internal}
