% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers_step4.R
\name{scbenchr_top_markers_per_cluster}
\alias{scbenchr_top_markers_per_cluster}
\title{Top markers per cluster helper}
\usage{
scbenchr_top_markers_per_cluster(dt, n = 10)
}
\arguments{
\item{dt}{data.table with cluster and avg_log2FC/avg_logFC column.}

\item{n}{number of genes per cluster.}
}
\value{
data.table of top n markers per cluster.
}
\description{
Top markers per cluster helper
}
