% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{scbenchr_write_session_info}
\alias{scbenchr_write_session_info}
\title{Write session info for reproducibility}
\usage{
scbenchr_write_session_info(out_dir, filename = "session_info.txt")
}
\arguments{
\item{out_dir}{Output directory.}

\item{filename}{Name of the session info file.}
}
\description{
Write session info for reproducibility
}
